requirements.txt
google-genai
pillow
python-dotenv
streamlit
openai
